minnor changes (snackbar and relocation of AutoFillNetworkUtils)
