+++ app/src/main/java/devs/erasmus/epills/widget/AddPillFinishDialog.java
@@ -5,14 +5,17 @@ import android.content.DialogInterface;
 import android.content.Intent;
 import android.os.Bundle;
 import android.support.annotation.NonNull;
+import android.support.design.widget.Snackbar;
 import android.support.v4.app.DialogFragment;
 import android.support.v7.app.AlertDialog;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+import android.widget.Toast;
 
 import devs.erasmus.epills.R;
 import devs.erasmus.epills.controller.AddPillSetTime;
+import devs.erasmus.epills.controller.MainActivity;
 
 
 /**
@@ -67,6 +70,7 @@ public class AddPillFinishDialog extends DialogFragment {
                                 activity.finish();
                                 break;
                             case 1:
+                                MainActivity.pillAddedSuccess();
                                 getActivity().finish();
                                 break;
                             default:
